Implemented subprotocol handling for both server and client
This is an optional feature requested by some WebSocket-based
clients such as noVNC or SPICE-HTML5. The list of supported
protocols can be specified by passing an array of strings under
the `:protocols` hash key. If no protocols are specified or an
empty array is passed, the server/client won't send the header.

The protocol array is required even for the hixie drafts that
have no support for multiple subprotocols. In such case the
first element of the array is used for comparison.

References:
https://tools.ietf.org/html/rfc6455#section-11.3.4
https://tools.ietf.org/html/draft-hixie-thewebsocketprotocol-75#section-8.5
https://tools.ietf.org/html/draft-hixie-thewebsocketprotocol-76#section-8.7